requirements.txt
streamlit==1.29.0
pypdf==4.0.1
google-generativeai==0.3.2
langchain==0.1.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
pyyaml==6.0.1